Look, the trees are turning their own bodies into pillars
of light,
are giving off the rich
fragrance of cinnamon
and fulfillment,

the long tapers
of cattails
are bursting and floating away over
the blue shoulders

of the ponds,
and every pond,
no matter what its
name is, is

nameless now.
Every year
everything
I have ever learned

in my lifetime
leads back to this: the fires
and the black river of loss
whose other side

is salvation,
whose meaning
none of us will ever know.
To live in this world

you must be able
to do three things:
to love what is mortal;
to hold it

against your bones knowing
your own life depends on it;
and, when the time comes to let it
go,
to let it go.

Mary Oliver
In Blackwater Woods